not use template

+++ math/prime.hpp
@@ -1,20 +1,19 @@
 #pragma once
 
-#include"../template/template.hpp"
-
+#include<utility>
 namespace noya2 {
 
-constexpr ll safe_mod(ll x, ll m) {
+constexpr long long safe_mod(long long x, long long m) {
     x %= m;
     if (x < 0) x += m;
     return x;
 }
 
-constexpr ll pow_mod_constexpr(ll x, ll n, int m) {
+constexpr long long pow_mod_constexpr(long long x, long long n, int m) {
     if (m == 1) return 0;
-    uint _m = (uint)(m);
-    ull r = 1;
-    ull y = safe_mod(x, m);
+    unsigned int _m = (unsigned int)(m);
+    unsigned long long r = 1;
+    unsigned long long y = safe_mod(x, m);
     while (n) {
         if (n & 1) r = (r * y) % _m;
         y = (y * y) % _m;
@@ -27,12 +26,12 @@ constexpr bool is_prime_constexpr(int n) {
     if (n <= 1) return false;
     if (n == 2 || n == 7 || n == 61) return true;
     if (n % 2 == 0) return false;
-    ll d = n - 1;
+    long long d = n - 1;
     while (d % 2 == 0) d /= 2;
-    constexpr ll bases[3] = {2, 7, 61};
-    for (ll a : bases) {
-        ll t = d;
-        ll y = pow_mod_constexpr(a, t, n);
+    constexpr long long bases[3] = {2, 7, 61};
+    for (long long a : bases) {
+        long long t = d;
+        long long y = pow_mod_constexpr(a, t, n);
         while (t != n - 1 && y != 1 && y != n - 1) {
             y = y * y % n;
             t <<= 1;
@@ -76,7 +75,7 @@ constexpr int primitive_root_constexpr(int m) {
     int cnt = 1;
     int x = (m - 1) / 2;
     while (x % 2 == 0) x /= 2;
-    for (int i = 3; (ll)(i)*i <= x; i += 2) {
+    for (int i = 3; (long long)(i)*i <= x; i += 2) {
         if (x % i == 0) {
             divs[cnt++] = i;
             while (x % i == 0) {
